+++ libs/src/ui/element/Toast/Toast.vue
@@ -1,105 +1,51 @@
 <script setup>
-if (!document.getElementById("toast")) {
-	const toastContainer = document.createElement("div");
-	toastContainer.id = "toast";
-	toastContainer.classList.add("ded-toast-container");
-	document.body.appendChild(toastContainer);
-}
+import Icon from "@/ui/element/Icon/Icon.vue";
+import Button from "@/ui/element/Button/Button.vue";
+import Title from "@/ui/element/Title/Title.vue";
+import { useToast } from "@/composable/useToast.js";
 
-import { ref, onMounted, onUnmounted } from 'vue';
-import Icon from '@/ui/element/Icon/Icon.vue';
-import Button from '@/ui/element/Button/Button.vue';
-import Title from '@/ui/element/Title/Title.vue';
-
-// 定義 Emits
-const emit = defineEmits(['onClose']);
+const { toasts, remove, positions } = useToast();
 
 // 定義 Props
 const props = defineProps({
-    themeColor: {
-        type: String,
-        validator: (value) =>
-            [
-                "primary",
-                "secondary",
-                "neutral",
-                "info",
-                "success",
-                "warning",
-                "error",
-            ].includes(value),
-        default: "primary",
-    },
-	title: {
-		type: String,
-		required: true
-	},
-	content: {
-		type: String,
-        default: "Content",
-	},
-    prefix: {
-        type: String,
-        default: "",
-    },
-	duration: {
-		type: Number,
-		default: 1000,
-	},
 	className: {
 		type: String,
 		default: "",
 	},
 });
 
-
-// 控制 關閉
-const handleClose = () => {
-	emit('onClose');
-};
-
-// 控制 自動關閉
-let timer = ref(null);
-const show = () => {
-	if (timer.value) clearTimeout(timer.value);
-	timer.value = setTimeout(() => {
-		handleClose();
-	}, props.duration);
-};
-
-// 元件掛載時 setTimeout
-onMounted(() => {
-	show();
-});
-
-// 元件卸載時卸載 setTimeout
-onUnmounted(() => {
-	if (timer.value) clearTimeout(timer.value);
-});
 </script>
 
 <template>
-	<teleport to="#toast">
-		<div class="ded-toast" :class="[`ded-toast-border-${props.themeColor}`, props.className]">
-			<!-- toast - 關閉按鈕 -->
-			<Button class="ded-close-button" variant="text" themeColor="neutral" @click="handleClose">
-				<Icon name="SvgClose" size="20"></Icon>
-			</Button>
-			<!-- toast - 標題及說明文字 -->
-            <div class="ded-toast-header">
-                <div :class="['ded-toast-header-message', `ded-toast-header-message-${props.themeColor}`]">
-                    <Icon :name="props.prefix" size="20"></Icon>
-                    <Title :themeColor="props.themeColor" :level="5">
-                        {{ props.title }}
-                    </Title>
+    <div v-for="position in positions" :key="position">
+        <div v-for="toast in toasts[position]" :key="toast.id">
+            <teleport :to="`#toast-container-${toast.position}`">
+                <div class="ded-toast" :class="[`ded-toast-border-${toast.themeColor}`, props.className]">
+                    <Button class="ded-close-button" variant="text" themeColor="neutral" @click="remove(toast.id, position)">
+                        <Icon name="SvgClose" size="20"></Icon>
+                    </Button>
+
+                    <div class="ded-toast-header">
+                        <div :class="['ded-toast-header-message', `ded-toast-header-message-${toast.themeColor}`]">
+                            <Icon :name="toast.prefix" size="20"></Icon>
+                            <Title :themeColor="toast.themeColor" :level="5">
+                                {{ toast.title }}
+                            </Title>
+                        </div>
+                        <div v-if="toast.action" class="ded-toast-header-action">
+                            <component v-if="typeof toast.action === 'object'" :is="toast.action"></component>
+                            <div v-else-if="typeof toast.action === 'function'">
+                                <component :is="toast.action()" />
+                            </div>
+                        </div>
+                    </div>
+
+                    <p class="ded-description">{{ toast.content }}</p>
                 </div>
-                <div class="ded-toast-header-action">
-                    <slot name="action"></slot>
-                </div>
-            </div>
-			<p class="ded-description">{{ props.content }}</p>
-		</div>
-	</teleport>
+            </teleport>
+        </div>
+    </div>
+
 </template>
 
 <style scoped>