+++ libs/src/ChangeLog.md
@@ -1,5 +1,23 @@
-## UAT V1.0.0 (2025-03-08)
+## DEV V1.0.0 (2025-03-13)
 
-### 🆙 Updates: (@ded-wds-vue/ui)
-- 因應 icon 終於測試成功，調整 Installation.md main.js 描述 
-- 整理沒用到的 .vue 與 .scss
+### 🆙 Updates: @ded-wds-vue/ui (學 #105)
+- 刪除 SvgMore.svg
+- 修改 icons.js import export
+- 同步學長 SCSS
+- Checkbox 1.加入唯一識別碼 2.在 Input 綁定 $attrs 接收 props 以外的屬性
+- Input 1.加入唯一識別碼 2.在 Input 綁定 $attrs 接收 props 以外的屬性
+- Radio 1.加入唯一識別碼 2.在 Input 綁定 $attrs 接收 props 以外的屬性
+- Table 1.修改 storybook
+- Textarea 1.加入唯一識別碼 2.在 Input 綁定 $attrs 接收 props 以外的屬性
+- ListItem 1.emits value 改 label
+- StatusIndicator 1.增加 isShowDot 屬性 2. variant 屬性增加 soft
+- Toast 1.改由 composable 控制
+- SideNave 1.修改 Navbar class
+
+### 🆙 Updates: @ded-wds-vue/ui (學 #106)
+- 同步學長 SCSS
+- 修改 storybook VNode
+
+### 🆙 Updates: @ded-wds-vue/ui (Amos 測試中)
+- 更新安裝指南文字
+- Icon 刪除 base64 讀取方法